[Minor] Types for Radio component
